fix!: Move execute to Ahoy instead of instruction
